+++ sunpy/database/database.py
@@ -380,6 +380,7 @@ class Database(object):
         path = kwargs.pop('path', None)
         progress = kwargs.pop('progress', False)
         methods = kwargs.pop('methods', ('URL-FILE_Rice', 'URL-FILE'))
+        caching = kwargs.pop('caching', True)
 
         if kwargs:
             k, v = kwargs.popitem()
@@ -388,7 +389,22 @@ class Database(object):
         if client is None:
             client = VSOClient()
 
-        paths = client.get(query_result, path, methods).wait(progress=progress)
+        if caching is True:
+            remove_list = []
+            for qr in query_result:
+                temp =  tables.DatabaseEntry._from_query_result_block(qr)
+                for database_entry in self:
+                    if database_entry.path is not None and temp._compare_attributes(
+                        database_entry, ["source", "provider", "physobs", "fileid", 
+                        "observation_time_start", "observation_time_end", 
+                        "instrument", "size", "wavemin", "wavemax"]):
+                        remove_list.append(qr)
+                        break
+
+            for temp in remove_list:
+                query_result.remove(temp)
+
+        paths = client.get(query_result, path).wait(progress=progress)
 
         for (path, block) in zip(paths, query_result):
             qr_entry = tables.DatabaseEntry._from_query_result_block(block)
@@ -426,6 +442,16 @@ class Database(object):
         from the VSO query result and the one from the downloaded FITS files)
         is added to the database in a way that each FITS header is represented
         by one database entry.
+
+        It uses the
+        :meth:`sunpy.database.Database._download_and_collect_entries` method
+        to download files, which uses query result block level caching. This
+        means that files will not be downloaded for any query result block
+        that had its files downloaded previously. If files for Query A were
+        already downloaded, and then a Query B is made which has some result
+        blocks common with Query A, then files for these common blocks will
+        not be downloaded again. Files will only be downloaded for those
+        blocks which are new or haven't had their files downloaded yet.
 
         """
         if not query:
@@ -742,13 +768,14 @@ class Database(object):
         self.add_from_vso_query_result(vso_qr, ignore_already_added)
 
     def download_from_vso_query_result(self, query_result, client=None,
-            path=None, progress=False, ignore_already_added=False):
+            path=None, progress=False, ignore_already_added=False, caching=True):
         """download(query_result, client=sunpy.net.vso.VSOClient(),
         path=None, progress=False, ignore_already_added=False)
 
         Add new database entries from a VSO query result and download the
         corresponding data files. See :meth:`sunpy.database.Database.download`
-        for information about the parameters `client`, `path`, `progress`.
+        for information about the caching mechanism used and about the
+        parameters `client`, `path`, `progress`.
 
         Parameters
         ----------
@@ -763,7 +790,7 @@ class Database(object):
         if not query_result:
             return
         self.add_many(self._download_and_collect_entries(
-            query_result, client=client, path=path, progress=progress))
+            query_result, client=client, path=path, progress=progress, caching=caching))
 
     def add_from_vso_query_result(self, query_result,
             ignore_already_added=False):